2293: - added methods to PluginFeatures interface - refactoring of Plugin logic from PluginBasedCommandlet to plugin manager
